Enable customization of localName for clients
This change enables custom LocalName entries to be specified for all
mails to be sent.

The motive is that smtp-relay.gmail.com heavily throttles the number of emails
that are accepted with a localName of "localhost". To get past the ehlo step in
this case the connection must specify another localName. The smtp library has
"localhost" hardcoded, so this change also inlines the default Send(...)
implementation.